Separate styles from auth components
