+++ src/components/SignUp/index.js
@@ -1,0 +1,3 @@
+import SignUp from "./SignUp";
+
+export default SignUp;
